--- a/assets/src/edit-story/components/checklist/checkpointContext/index.js
+++ b/assets/src/edit-story/components/checklist/checkpointContext/index.js
@@ -17,4 +17,5 @@
 export {
   ChecklistCheckpointProvider,
   useCheckpoint,
+  CheckpointContext,
 } from './checkpointContext';