--- a/assets/src/edit-story/components/checklist/checks/imageElementMissingAlt.js
+++ b/assets/src/edit-story/components/checklist/checks/imageElementMissingAlt.js
@@ -48,8 +48,9 @@ const ImageElementMissingAlt = () => {
   const elements = filterStoryElements(story, imageElementMissingAlt);
   const setHighlights = useHighlights(({ setHighlights }) => setHighlights);
   const handleClick = useCallback(
-    (elementId) =>
+    (elementId, pageId) =>
       setHighlights({
+        pageId,
         elementId,
         highlight: states.ASSISTIVE_TEXT,
       }),
@@ -76,7 +77,7 @@ const ImageElementMissingAlt = () => {
             {getVisibleThumbnails(elements).map((element) => (
               <Thumbnail
                 key={element.id}
-                onClick={() => handleClick(element.id)}
+                onClick={() => handleClick(element.id, element.pageId)}
                 type={THUMBNAIL_TYPES.IMAGE}
                 displayBackground={<LayerThumbnail page={element} />}
                 aria-label={__('Go to offending video', 'web-stories')}