--- a/assets/src/edit-story/components/checklist/checks/imageElementResolution.js
+++ b/assets/src/edit-story/components/checklist/checks/imageElementResolution.js
@@ -74,8 +74,9 @@ const ImageElementResolution = () => {
   );
   const setHighlights = useHighlights(({ setHighlights }) => setHighlights);
   const handleClick = useCallback(
-    (elementId) =>
+    (elementId, pageId) =>
       setHighlights({
+        pageId,
         elementId,
       }),
     [setHighlights]
@@ -107,7 +108,7 @@ const ImageElementResolution = () => {
             {getVisibleThumbnails(failingElements).map((element) => (
               <Thumbnail
                 key={element.id}
-                onClick={() => handleClick(element.id)}
+                onClick={() => handleClick(element.id, element.pageId)}
                 type={THUMBNAIL_TYPES.IMAGE}
                 displayBackground={<LayerThumbnail page={element} />}
                 aria-label={__('Go to offending image', 'web-stories')}