--- a/assets/src/edit-story/components/checklist/checks/textElementFontSizeTooSmall.js
+++ b/assets/src/edit-story/components/checklist/checks/textElementFontSizeTooSmall.js
@@ -48,9 +48,10 @@ const TextElementFontSizeTooSmall = () => {
   const elements = filterStoryElements(story, textElementFontSizeTooSmall);
   const setHighlights = useHighlights(({ setHighlights }) => setHighlights);
   const handleClick = useCallback(
-    (elementId) =>
+    (elementId, pageId) =>
       setHighlights({
         elementId,
+        pageId,
       }),
     [setHighlights]
   );
@@ -74,7 +75,7 @@ const TextElementFontSizeTooSmall = () => {
             {getVisibleThumbnails(elements).map((element) => (
               <Thumbnail
                 key={element.id}
-                onClick={() => handleClick(element.id)}
+                onClick={() => handleClick(element.id, element.pageId)}
                 type={THUMBNAIL_TYPES.TEXT}
                 displayBackground={<LayerThumbnail page={element} />}
                 aria-label={__('Go to offending text element', 'web-stories')}