--- a/assets/src/edit-story/components/checklist/checks/videoElementLength.js
+++ b/assets/src/edit-story/components/checklist/checks/videoElementLength.js
@@ -48,9 +48,10 @@ const VideoElementLength = () => {
   const failingElements = filterStoryElements(story, videoElementLength);
   const setHighlights = useHighlights(({ setHighlights }) => setHighlights);
   const handleClick = useCallback(
-    (elementId) =>
+    (elementId, pageId) =>
       setHighlights({
         elementId,
+        pageId,
       }),
     [setHighlights]
   );
@@ -75,7 +76,7 @@ const VideoElementLength = () => {
             {getVisibleThumbnails(failingElements).map((element) => (
               <Thumbnail
                 key={element.id}
-                onClick={() => handleClick(element.id)}
+                onClick={() => handleClick(element.id, element.pageId)}
                 type={THUMBNAIL_TYPES.VIDEO}
                 displayBackground={<LayerThumbnail page={element} />}
                 aria-label={__('Go to offending video', 'web-stories')}