… assets/src/edit-story/app/prepublish/components/videoOptimization.js
… assets/src/edit-story/app/prepublish/constants.js
… assets/src/edit-story/app/prepublish/error/index.js
… assets/src/edit-story/app/prepublish/error/metadata.js
… assets/src/edit-story/app/prepublish/error/test/metadata.js
… assets/src/edit-story/app/prepublish/getPrepublishErrors.js
… assets/src/edit-story/app/prepublish/guidance/general.js
… assets/src/edit-story/app/prepublish/guidance/index.js
… assets/src/edit-story/app/prepublish/guidance/media.js
… assets/src/edit-story/app/prepublish/guidance/test/general.js
… assets/src/edit-story/app/prepublish/guidance/test/media.js
… assets/src/edit-story/app/prepublish/guidance/test/text.js
… assets/src/edit-story/app/prepublish/guidance/text.js
… assets/src/edit-story/app/prepublish/test/__snapshots__/getPrepublishErrors.js.snap
… assets/src/edit-story/app/prepublish/test/getPrepublishErrors.js
… assets/src/edit-story/app/prepublish/types.js
… assets/src/edit-story/app/prepublish/warning/accessibility.js
… assets/src/edit-story/app/prepublish/warning/distribution.js
… assets/src/edit-story/app/prepublish/warning/index.js
… assets/src/edit-story/app/prepublish/warning/test/accessibility.js
… assets/src/edit-story/app/prepublish/warning/test/distribution.js
… assets/src/edit-story/components/checklist/stories/reviewChecklistDialog.js
… assets/src/edit-story/components/header/buttons/buttonWithChecklistWarning.js
… assets/src/edit-story/components/header/buttons/publish.js
… assets/src/edit-story/components/header/buttons/switchToDraft.js
… assets/src/edit-story/components/header/test/buttons.js
… assets/src/edit-story/components/inspector/constants.js
… assets/src/edit-story/components/inspector/index.js
… assets/src/edit-story/components/inspector/inspectorProvider.js
… assets/src/edit-story/components/inspector/karma/checklistTab.karma.js