## Context

The final bit of redoing the checklist to be outside of the inspector panel. 

## Summary

Removes the old checklist (that was in the inspector panel) in favor of the new version that is in a popup. Sets the new checklist experiment to true by default. 

Figma: https://www.figma.com/file/1YM3jA9h4QGc4xLdyaXIJS/Stories-Sprint-1.7?node-id=8378%3A3778 

## Relevant Technical Choices

Mostly I deleted code.  
- Rewired the header buttons to use just the new implementation.
- Because of improvements made we don't need to watch anything on the draft button to reset the checklist to watch for the dialog 
- Moved the review dialog and storybook for that and the checklist icon that shows up in the 'publish' button into the new `components/checklist` 
- Copied over the useful karma tests in the directories I deleted and updated them to use our new structure
- These tests helped me catch some regressions where I just didn't understand the highlight hook super well last week, so I fixed those callbacks and got the tests working again. 

## To-do

- [x] Screen casts of updates 
- [x] In-depth testing directions since this will be Claudio's QA ticket 

## User-facing changes

- No more checklist in inspector panel 
- Checklist exists next to help center as a pop up 

| Before | After | 
| --- | --- |
| ![Screen Shot 2021-07-02 at 1 03 38 PM](https://user-images.githubusercontent.com/10720454/124323294-f11ea880-db35-11eb-94be-eb55dfb0ea70.png) | ![Screen Shot 2021-07-02 at 1 01 54 PM](https://user-images.githubusercontent.com/10720454/124323196-bfa5dd00-db35-11eb-903f-7e54bbd14d9c.png) |

| Little demo | 
| --- | 
| https://user-images.githubusercontent.com/10720454/124325580-c33b6300-db39-11eb-9b7c-e7acb84b09c0.mp4 |



## Testing Instructions

#### Checklist Pop up 
- The pop up should open and close by hitting the toggle button for it (looks like a checkmark kinda, next to the help center in the same row as the carousel) 
- It should close when you hit the 'x' in the popup header 
- The count in the toggle button should update to show you how many issues you have. 
- To begin with no count should show. Only when checklist categories begin to show up will the count also show up. 
- Please also make sure that the help center still works as intended

#### Publishing 
- If there are 'priority' issues in the checklist (this isn't new), then an alert icon shows next to the publish button. 
- If you click publish when there are priority issues you will trigger a dialog asking if you want to review the checklist (not new). 
- If you choose to review the checklist, regardless of how many pages/elements you have added to the story you will now have all sections of the checklist exposed (3 total, but accessibility doesn't show up on a blank 1 page story) and the priority section should be expanded by default. 
- If you choose to publish anyways it should just publish your story like normal, now that the story is published the checklist is visible in its entirety even if you publish a blank 1 page story. 

#### Checkpoints 
- There are several 'checkpoints' you hit while creating a story that enable sections of the checklist. 
##### Checkpoint 1: Recommended Sections 
- These include design and accessibility
- They will be triggered when you add any element that is not a background color or a second page 

##### Checkpoint 2: Priority 
- This will be triggered when you've added a 5th page 
- OR you jump the gun and publish anytime before then. 
- Once priority is visible so is design and accessibility - if you skipped over those.
 
#### Checklist Categories 
- There are now 3 categories: priority, design, accessibility 
- Only 1 category panel is expanded at a time
- The panel you have expanded should be retained on next expansion 
- The only exception to this is if you trigger a checkpoint (see above).
- If you have no issues to resolve in a category, that category will not show up. 

#### All of the Checks 🥵  
- There's nothing new here! We want to avoid regression from these checks that are also present in the old checklist. [This is a list of the checks](https://docs.google.com/spreadsheets/d/1PjaRp4bueQ3Cv14A7I696m0vjGCIX_kN9tIGh-V-Bkc/edit#gid=920793050) to help assist with QA and making sure we didn't miss anything

| Check | Section | Interactivity |
| --- | --- | --- |
| Font size too small | a11y | Has a thumbnail you can click, if you do you go to the text element |
| Images missing alt text | a11y | Has a thumbnail you can click, if you do you go to the alt text input |
| Link tappable region too small | a11y | Has a thumbnail you can click, if you do you go to the offending element |
| Low contrast between font and bg color | a11y | Has a thumbnail you can click, if you do you go to the offending element |
| Video doesn't include captions | a11y | Has a thumbnail you can click, if you do you go to the video captions upload | 
| Video missing title | a11y | Has a thumbnail you can click, if you do you go to the video title input | 
| Story length is bad ( less than 4 more than 30) | Design | n/a |
| Too little text | Design | n/a |
| Too much text on a page | Design | Has a thumbnail that will take you to the page with all the text, all of the text elements on that page will be selected | 
| Too many links on a page | Design | Has a thumbnail that will take you to the page with all the text, all of the text elements on that page will be selected | 
| Low image resolution poster image | Design | Has a thumbnail, take you to the offending element | 
| video less than 24 fps | Design |  Has a thumbnail, takes you to the offending element | 
| video resolution too high | Design | Has a thumbnail, takes you to the offending element  |
| video is long | Design |  Has a thumbnail, takes you to the offending element | 
| Missing story description | Priority | Click the title area, takes you to story description input | 
| Missing title | Priority | Click title area, takes you to story title input | 
| Missing story portrait poster | Priority | Takes you to upload for poster |
| Video is missing poster | Priority | Takes you to offending element and upload for poster for that video | 
| Story title is longer than 70 char | Priority | Click title area, takes you to story title input | 
| Publisher logo too small | Priority | Click title area, go to publisher logo upload button | 
| story portrait too small | Priority | Click title area or thumbnail and go to publisher logo upload button | 
| Story portrait wrong aspect ratio | Priority | Click title area or thumbnail and go to publisher logo upload button | 
| Video not optimized | Priority |  If optimization is enabled then you will get a card with a thumbnail and a button to optimize your video. It's pretty cool. If optimization is not enabled, this doesn't work. Instead, once the accessibility section is visible, you get an ask if you want to enable optimization for the future |


### QA

<!--
Not all changes require manual QA.
-->

<!-- ignore-task-list-start -->
- [ ] This is a non-user-facing change and requires no QA
<!-- ignore-task-list-end -->

This PR can be tested by following these steps:

1. See the above :) 

### UAT

<!--
Sometimes the testing instructions for UAT can differ from the ones for QA.
-->

<!-- ignore-task-list-start -->
- [x] UAT should use the same steps as above.
<!-- ignore-task-list-end -->

<!--
If the above checkbox has not been checked, write down all steps necessary for user acceptance testing take to test this PR.
-->
This PR can be tested by following these steps:

1.

## Reviews

### Does this PR have a security-related impact?

No

### Does this PR change what data or activity we track or use?

No

### Does this PR have a legal-related impact?

No

## Checklist

<!-- Check these after PR creation -->

- [x] This PR addresses an existing issue and I have linked this PR to it in ZenHub
- [x] I have tested this code to the best of my abilities
- [x] I have verified accessibility to the best of my abilities ([docs](https://github.com/google/web-stories-wp/blob/main/docs/accessibility-testing.md))
- [x] I have verified i18n and l10n (translation, right-to-left layout) to the best of my abilities
- [x] This PR contains automated tests (unit, integration, and/or e2e) to verify the code works as intended ([docs](https://github.com/google/web-stories-wp/tree/main/docs#testing))
- [x] I have added documentation where necessary
- [x] I have added a matching `Type: XYZ` label to the PR

---

<!--
Please reference the issue(s) this PR addresses.
No URLs, just the issue numbers.
Use "Fixes #123" if it fixes an issue.
Example:

Fixes #123
Partially addresses #456
See #789
-->

Fixes #7978