@samwhale 2021-07-02
Note: the checklist is still visible to screen readers, but that should be fixed in #8172. Once that's merged we won't have this problem:
<img src="https://user-images.githubusercontent.com/22185279/124328961-d2291200-db47-11eb-961f-81857fcef75f.gif" height="375px" />