@samwhale 2021-07-05
@googlebot I consent.