@samwhale 2021-07-05
> Have you looked into why Editor Integration Tests are failing?

Yeah just fixed this - the tablist isn't visible anymore when the checklist is closed so we can't select for `data-isexpanded='false'`. Removing that check works since we still check for existence in other ways