@samwhale 2021-07-06
These two tests keep failing, but at first glance seem unrelated to the changes we made. Will check tomorrow:

<img src="https://user-images.githubusercontent.com/22185279/124539930-a8881900-dddb-11eb-81c7-9d5ee4a5538c.png" height="300px" /> <img src="https://user-images.githubusercontent.com/22185279/124539941-acb43680-dddb-11eb-9666-f75e2350f8ec.png" width="300px" />