@BrittanyIRL 2021-07-06
> > User-facing changes
> 
> Should the Design & Document tabs now perhaps be more evenly spaced?

This is how design wanted it: https://www.figma.com/file/1YM3jA9h4QGc4xLdyaXIJS/Stories-Sprint-1.7?node-id=8378%3A0

![Screen Shot 2021-07-06 at 8 42 36 AM](https://user-images.githubusercontent.com/10720454/124629350-34c72a00-de36-11eb-809f-ee2cb22109b1.png)