@BrittanyIRL 2021-07-06
> I think the badge next to the ppc companion toggle is only counting priority issues. Is this an intentional request from design?

Great question! This was intentional. It's a carry over from the same structure as while the checklist was in the inspector panel. Badge shows up when priority is visible as a panel and only counts the priority issues..