@spacedmonkey 2021-07-06
> These two tests keep failing, but at first glance seem unrelated to the changes we made. Will check tomorrow:

Known issue https://github.com/google/web-stories-wp/issues/8160 @samwhale . I will work on it tomorrow.