@BrittanyIRL 2021-07-07
> The PPC e2e tests are failing. These need update to work again.
> 
> https://github.com/google/web-stories-wp/blob/main/packages/e2e-tests/src/specs/editor/prepublishChecklist.js

@spacedmonkey - tests are all merged in so this branch is passing now. Thanks for the quick fix on the e2e tests that you merged earlier today.